+++ src/azure/identity/default_azure_credential.rs
