+++ src/azure/identity/mod.rs
@@ -1,0 +1,5 @@
+mod default_azure_credential;
+mod python_azure_credential;
+
+pub(crate) use default_azure_credential::DefaultAzureCredential;
+pub use python_azure_credential::PythonAzureCredential;
