+++ src/azure/key_vault/mod.rs
@@ -1,10 +1,7 @@
 mod azure_key_vault_settings_provider;
 mod azure_key_vault_settings_source;
-mod default_azure_credential;
 mod parallel_secret_loader;
-mod python_azure_credential;
 mod remove_user_agent;
 
 pub use azure_key_vault_settings_provider::AzureKeyVaultSettingsProvider;
 pub use azure_key_vault_settings_source::AzureKeyVaultSettingsSource;
-pub use python_azure_credential::PythonAzureCredential;
