+++ src/azure/key_vault/parallel_secret_loader.rs
