+++ src/azure/mod.rs
@@ -1,0 +1,2 @@
+pub mod identity;
+pub mod key_vault;
